Removed encode-unicode dependency.
The reason for attempting to remove this is that encode-unicode provides
some additional iterators that can make some type inferences break by
including it (or a crate that depends on it, like prettytable-rs).

For example, the following line of code works as expected if you include
the `rand` crate:

```
use rand::thread_rng;
use rand::distributions::{Distribution, Standard};

fn main() {
  let mut rng = thread_rng();
  let v: Vec<u8> = Standard.sample_iter(&mut rng).take(5).collect();
  println!("{:}", v);
}
```

However, once prettytable-rs is included and used, the types will no
longer be able to be inferred.

```
error[E0282]: type annotations needed
  --> src/main.rs:16:31
   |
16 |     let v: Vec<u8> = Standard.sample_iter(&mut rng).take(5).collect();
   |                               ^^^^^^^^^^^
   |                               |
   |                               cannot infer type for type parameter `T`
```

While not technically a problem that NEEDS to be fixed (type inference
breaks are not really considered "breaking changes, it can be a pain
point. Seeing that prettytable-rs doesn't seem to make much use of the
additional items in encode-unicode, perhaps it's not worth the pain to
users to include it.